EP PyTorch: move symm_mem_alloc to distributed.py and re-export from ep
