+++ transformer_engine/pytorch/distributed.py
@@ -1842,6 +1842,27 @@ def get_symmetric_memory_tensor(tensor_numel, tensor_dtype, tensor_device, tp_gr
     return msg
 
 
+def symm_mem_alloc(
+    shape,
+    dtype: torch.dtype,
+    ep_group: dist_group_type,
+    device: Optional[torch.device] = None,
+) -> torch.Tensor:
+    """Allocate and rendezvous a symm-mem buffer on ep_group. Collective on ep_group."""
+    if device is None:
+        device = torch.device("cuda", torch.cuda.current_device())
+    if not HAS_TORCH_SYMMETRIC:
+        raise RuntimeError(
+            "torch.distributed._symmetric_memory is unavailable; symm_mem_alloc "
+            "requires PyTorch built with NCCL symm-mem support."
+        )
+    if symm_mem.get_backend(device) != "NCCL":
+        symm_mem.set_backend("NCCL")
+    t = symm_mem.empty(*shape, dtype=dtype, device=device)
+    symm_mem.rendezvous(t, group=ep_group.group_name)
+    return t
+
+
 def symmetric_all_reduce(
     inp: torch.Tensor,
     tp_group: Optional[dist_group_type] = None,
